+++ app/ExternalFiles/Protocol_Base.php
@@ -158,7 +158,7 @@ class Protocol_Base {
 	}
 
 	/**
-	 * Check the availability of a given URL.
+	 * Return infos to each given URL.
 	 *
 	 * @return array List of files with its infos.
 	 */