+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Added Services to platforms which host files, usable for local host, FTP, YouTube and ZIP-files atm
 - Added management for credentials of these external platforms
+- Added possibility to import files from Google Drive as external files
 - Added possibility to add YouTube- and Vimeo-videos as external files
 - Added possibility to import Youtube-Channel-Videos as external files via YouTube API
 - Added marker for proxied files